+++ src/utils/d3-linear-nice.ts
@@ -5,9 +5,17 @@ import { tickIncrement } from './ticks';
 import { NiceMethod } from '../types';
 
 export const d3LinearNice: NiceMethod = (min: number, max: number, count: number = 5) => {
-  let start = min < max ? min : max;
-  let stop = min < max ? max : min;
-  let step: number;
+  const d = [min, max];
+  let i0 = 0;
+  let i1 = d.length - 1;
+  let start = d[i0];
+  let stop = d[i1];
+  let step;
+
+  if (stop < start) {
+    [start, stop] = [stop, start];
+    [i0, i1] = [i1, i0];
+  }
   step = tickIncrement(start, stop, count);
 
   if (step > 0) {
@@ -21,11 +29,11 @@ export const d3LinearNice: NiceMethod = (min: number, max: number, count: number
   }
 
   if (step > 0) {
-    start = Math.floor(start / step) * step;
-    stop = Math.ceil(stop / step) * step;
+    d[i0] = Math.floor(start / step) * step;
+    d[i1] = Math.ceil(stop / step) * step;
   } else if (step < 0) {
-    start = Math.ceil(start * step) / step;
-    stop = Math.floor(stop * step) / step;
+    d[i0] = Math.ceil(start * step) / step;
+    d[i1] = Math.floor(stop * step) / step;
   }
-  return [start, stop];
+  return d;
 };
