+++ src/utils/index.ts
@@ -7,6 +7,7 @@ export { d3LinearNice } from './d3-linear-nice';
 export { d3TimeNice } from './d3-time-nice';
 export { isValid } from './is-valid';
 export { tickIncrement, tickStep } from './ticks';
+export { findTickInterval } from './find-tick-interval';
 
 export {
   DURATION_SECOND,
